task 16. In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,24 +6,23 @@
  * Return: 0
  */
 
-
 int main(void)
 {
-	int n1 = 1, n2 = 2, next;
+	unsigned long int n1 = 1, n2 = 2, next;
 	int count;
 
-	printf("%d, %d, ", n1, n2);
+	printf("%lu, %lu, ", n1, n2);
 
 	for (count = 3; count <= 98; count++)
 	{
 		next = n1 + n2;
 		if (count < 98)
 		{
-			printf("%d, ", next);
+			printf("%lu, ", next);
 		}
 		else
 		{
-			printf("%d\n", next);
+			printf("%lu\n", next);
 		}
 		n1 = n2;
 		n2 = next;
